split methods for rise and kernelshap
