+++ dianna/methods/kernelshap_image.py
