+++ dianna/utils/predict.py
@@ -1,3 +1,4 @@
+"""Utility functions for perturbation based predictions."""
 import numpy as np
 from tqdm import tqdm
 
@@ -20,3 +21,7 @@ def make_predictions(data, runner, batch_size):
     for i in tqdm(range(0, number_of_masks, batch_size), desc="Explaining"):
         predictions.append(runner(data[i : i + batch_size]))
     return np.concatenate(predictions)
+
+def normalize(saliency, n_masks, p_keep):
+    """Normalizes salience by number of masks and keep probability."""
+    return saliency / n_masks / p_keep
